refactor: consolidated logic and extension files to controller
