+++ lib/features/history/ui/screen/workout_history_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flex_workout_mobile/core/extensions/ui_extensions.dart';
 import 'package:flex_workout_mobile/core/theme/app_layout.dart';
-import 'package:flex_workout_mobile/features/tracker/ui/containers/history_calendar.dart';
-import 'package:flex_workout_mobile/features/tracker/ui/containers/workout_history_list.dart';
+import 'package:flex_workout_mobile/features/history/ui/containers/history_calendar.dart';
+import 'package:flex_workout_mobile/features/history/ui/containers/workout_history_list.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
